+++ raster/src/main/scala/geotrellis/raster/io/geotiff/compression/Predictor.scala
@@ -16,14 +16,24 @@
 
 package geotrellis.raster.io.geotiff.compression
 
-import geotrellis.raster.io.geotiff.tags._
+import geotrellis.raster.io.geotiff.GeoTiffImageData
 import geotrellis.raster.io.geotiff.reader.MalformedGeoTiffException
+import geotrellis.raster.io.geotiff.tags._
+import geotrellis.raster.io.geotiff.tags.codes.SampleFormat._
 import monocle.syntax.apply._
 
 object Predictor {
   val PREDICTOR_NONE = 1
   val PREDICTOR_HORIZONTAL = 2
   val PREDICTOR_FLOATINGPOINT = 3
+
+  def apply(imageData: GeoTiffImageData): Predictor = {
+    imageData.bandType.sampleFormat match {
+      case SignedInt | UnsignedInt => HorizontalPredictor(imageData)
+      case FloatingPoint => FloatingPointPredictor(imageData)
+      case _ => throw new UnsupportedOperationException(s"Only predictor 2 (integer-based) and predictor 3 (float-based) are supported.")
+    }
+  }
 
   def apply(tiffTags: TiffTags): Predictor = {
     (tiffTags
@@ -32,9 +42,11 @@ object Predictor {
     ) match {
       case None | Some(PREDICTOR_NONE) =>
         new Predictor {
-          val code = PREDICTOR_NONE
+          val code: Int = PREDICTOR_NONE
           val checkEndian = true
-          def apply(bytes: Array[Byte], segmentIndex: Int) = bytes
+
+          def encode(bytes: Array[Byte], segmentIndex: Int): Array[Byte] = bytes
+          def decode(bytes: Array[Byte], segmentIndex: Int): Array[Byte] = bytes
         }
       case Some(PREDICTOR_HORIZONTAL) =>
         HorizontalPredictor(tiffTags)
@@ -44,6 +56,24 @@ object Predictor {
         throw new MalformedGeoTiffException(s"predictor tag $i is not valid (require 1, 2 or 3)")
     }
   }
+
+  def colsPerRow(imageData: GeoTiffImageData): Int =
+    if (imageData.segmentLayout.isStriped)
+      imageData.segmentLayout.totalCols
+    else
+      imageData.segmentLayout.tileLayout.tileCols
+
+  def rowsInSegment(imageData: GeoTiffImageData): Int => Int =
+    if (imageData.segmentLayout.isStriped) {
+      def stripedRowsInSegment(segmentIndex: Int): Int =
+        imageData.segmentLayout.getSegmentDimensions(segmentIndex).rows
+
+      stripedRowsInSegment
+    } else {
+      def tiledRowsInSegment(segmentIndex: Int): Int =
+        imageData.segmentLayout.tileLayout.tileRows
+      tiledRowsInSegment
+    }
 }
 
 trait Predictor extends Serializable {
@@ -52,5 +82,7 @@ trait Predictor extends Serializable {
   /** GeoTiff tag value for this predictor */
   def code: Int
 
-  def apply(bytes: Array[Byte], segmentIndex: Int): Array[Byte]
+  def encode(bytes: Array[Byte], segmentIndex: Int): Array[Byte]
+
+  def decode(bytes: Array[Byte], segmentIndex: Int): Array[Byte]
 }